requirements.txt
biopython==1.85
numpy==2.3.3
